requirements.txt
boto3==1.42.34
botocore==1.42.34
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
idna==3.11
jmespath==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==1.26.20
youtube-transcript-api==1.2.3
